Ref #16924: Reimplement enums
